Cleanup in UnreachableCodeElimination(Test)
Eliminate the need to run normalize on expected outputs in order to make the
test cases easier to understand.

General formatting cleanup for both the implementation class and the tests.

There are still a lot of badly formatted test cases.
It's just not worth my time to fix them right now.

PiperOrigin-RevId: 561182597