Fix go/ts52upgrade JSC_MISPLACED_ANNOTATION issue
For decorated classes, that reference the class name in a static block, TypeScript 5.2 changed the JS emit from

    /** @abstract */
    let Foo = class Foo {

to

    var Foo_1;
    /** @abstract */
    let Foo = Foo_1 = class Foo {

This caused JSC_MISPLACED_ANNOTATION errors for both `@abstract` and `@template`. The only fix needed is in the TypedScopeCreator.

Optimizations InlineAndCollapseProperties and RemoveUnusedCode already don't optimize the 5.2 emit. They work the same on 5.3 emit.

PiperOrigin-RevId: 561277685