UnreachableCodeElimination: Avoid traversing unreachable statements
When traversing a block of statements, skip traversal of all statements
following the first one we find to be unreachable and remove them all.

This avoids unnecessary work traversing the individual statements that we
already know will not ever execute.

This change also fixes a bug that caused unreachable `do {} while (condition)`
statements to be left behind. This could cause an undeclared variable reference
if the condition referred to a `let` or `const` variable whose declaration was
also removed as unreachable.

That bug was the original motivation for this change.

PiperOrigin-RevId: 565691433