Prepare for upgrade to java-diff-utils 4.12
The new version of the library doesn't have a `DiffException` that the
current version forces us to catch. We cannot upgrade to the library and
change the code that uses it in the same change for complicated reasons,
so instead catch `Exception` to remove the use of a soon-to-be-eliminated
class.

PiperOrigin-RevId: 567365082