Add more details of conformance violations to the JSON error report.
In the case of conformance violations, the JSON report that's generated with warnings (with the `json_extra_warnings_file` flag) contains the violations with just the error message, and no other information about the conformance rule -- the rule id, or the config where the rule is defined, etc.  In this CL, I am adding more relevant details of the conformance rule to the JSON file.

PiperOrigin-RevId: 580375320