Enable type checking and colors for J2clPassTest
It's important that the correct colors be applied to nodes that J2clPass
may create in order to avoid problems with PropertyDisambiguation/Ambiguation.
This change just enables the `CompilerTestCase` options that will ensure
that nodes have colors applied to them when the tests run.

Also add logic to one test to verify that the correct color is applied
when inlining a `$markImplementor()` call.

PiperOrigin-RevId: 581383647