Skip marking the for and for-in initializer as non-const when converting them to var
`for (const a in b) { use(a); }`

converts to

`var a; for (a in b) { use(a); };`

If the initializer name was a const, it must no longer be a marked const. However, marking it as non-const requires us to also find and change all references to the initializer name in the for-in's (or vanilla for's) body. That is expensive to do. Hence, we leave the initializer name and its references within the loop to be const.

This does not degrade optimizations as the references within the loop continue to be marked const currently even after normalize.

How does it work currently?

Currently, we rewrite first using ES6RewriteBlockScopedDeclarations to `var a; for (a in b) { use(a);};` and then run Normalize. Normalize pass does not mark the use(a) in the for-in body as const.

How will it work later?

When we will change to run ES6RewriteBlockScopedDeclarations after Normalize, the Normalize pass will run first and see `for (const a in b) { use(a)}` and mark both instances of `a` as const. With this CL we make the ES6RewriteBlockScopedDeclarations keep them both const to avoid expensive traversals for searching `use(a)`.

PiperOrigin-RevId: 584432318