Add unit test for collapse properties issue
TS 5.2 starts outputing code in a shape that causes a bad property collapse.

This change adds a unit test for InlineAndCollapseProperties that mimics
a previously-added reproduction of this problem in an integration test.

At the moment the unit test added by this change doesn't successfully reproduce
the issue, but I'd like to check it in anyway as a step in the investigation.

The next step is to figure out why a different decision is being made
by InlineAndCollapseProperties in this unit test vs the integration test case
that does reproduce the problem.

PiperOrigin-RevId: 588533168