Fix CrossChunkCodeMotion to consider computed props that are Symbols as movable.
Currently, JSCompiler only moves computed properties iff the value is guaranteed and known by JsCompiler to not change.

`Symbol` is a built-in JavaScript object that is guaranteed to be unique. Therefore it is movable, and this change tells CCCM that it should consider `Symbol` as moveable.

This change makes CCCM consider any properties on `Symbol` as movable (e.g `Symbol.iterator`), which should make this futureproof if the JavaScript spec changes.

More info on `Symbol` - go/mdn/JavaScript/Reference/Global_Objects/Symbol

PiperOrigin-RevId: 592686737