A better workaround for the LICENSE problem.
Don't use the deploy jar when creating the unshaded release jar, because
that effectively ceates a shaded jar with a risk of naming collisions.
Instead include the `LICENSE` file under 2 names to work around a
bug/feature in the `java_export` rule.

Also clean up some unnecessary rules and attributes that were only
present due to past attempts to fix this same problem.