Stop creating block scope for synthetic block nodes
A BLOCK marked as synthetic is not a real JS block with {} around it in the JS code. It just represents a span of statements that need to be kept together. Synthetic blocks do not exist in the source AST and hence do not create their own scope.

This CL impacts the NodeTraversal class and several JSCompiler passes.

PiperOrigin-RevId: 618383865