Fix potential property disambiguation bug for Polymer 1 code
The root cause was that the PolymerPass adds nodes to the synthetic externs file, but the disambiguator does not consistently recognize these nodes as actually being from the externs.

This a known problem in b/186056977 where the compiler doesn't enforce that synthetic extern nodes have property source info. I think the right fix is for the disambiguator to handle this case by always looking at the SCRIPT node to see if a node is in the externs or not.

PiperOrigin-RevId: 625696735