Adds `isZoneInput` and `isZoneEnabled` to configure how the compiler detects the presence of Zone.js.
Previously, Zone.js was detected by checking a google3-specific path and throwing an error when this was used in combination with ES2017+ output, since Zone.js is incompatible with native async/await. `isZoneInput` and `isZoneEnabled` allow configuring the compiler to treat any particular input and flag as controlling Zone.js. This allows JSCompiler to emit ES2017+ output as long as Zone.js is not included in the compilation, or is disabled and tree-shaken via a `goog.define` flag.

PiperOrigin-RevId: 629455485