Rewrite destructuring var declarations in Normalize to stub var declarations of individual names
#### Problem:
Normalize is inconsistent about rewriting duplicate declarations. It does not rewrite redeclaration like below in which one of the redeclaration is inside a destructuring syntax:

```
var a = 4; var [a] = [5];
```
OR

```
var a = 4; var {a: a} = {a: 5};
```

#### Why does it matter?

This becomes a problem when Es6RewriteDestructuring pass runs on this code post Normalize (and Normalize has failed to de-dup the 2 declares). In that case, the Es6RewriteDestructuring produces duplicate declarations (like below) thereby making the AST unnormalized again.

```
var a = 4; var a = 5;
```

#### What will get fixed?

In this CL, I am fixing Normalize to rewrite `var a = 4; var [a] = [5];` into `var a = 4; [a] = [5];` and rewrite `var a = 4; var {a: a} = {a: 5};` into `var a = 4; ({a:a} = {a: 5});`

So that Es6RewriteDestructuring can then generate `var a = 4; a = 5;` for the first example, and `var a = 4; a = 5;` for the second example (i.e no duplicate declarations and AST stays normalized).

PiperOrigin-RevId: 634956271