Add `$jscomp.patch` runtime function
This function behaves similarly to `$jscomp.polyfill` in terms of dead code removal, but is never injected based on usage.  The idea is that these function calls are injected based on other factors (i.e. polyfilling a related symbol) and can then be removed if the patched symbol is never used.

A follow-up CL will update the `WeakMap` and other patching polyfills to use this instead.  The primary motivation is `AsyncContext`, which needs to patch a significant number of rarely-used APIs, and we'd like to avoid the extra code size overhead when the patched APIs are unused.

PiperOrigin-RevId: 639825637