Add an `InstrumentAsyncContext` pass.
This pass instruments `async` functions and generators to add `$jscomp.asyncContextEnter`, `Exit`, and `Reenter` calls around the async entry/exit points.  The associated runtime library is initially a no-op, but it will later be patched by the `AsyncContext.Variable` polyfill.

This pass is off by default for now.  A future flag flip CL will go in once a global presubmit is clean.

PiperOrigin-RevId: 641953935