Add `AsyncContext.Variable` polyfill.
This polyfill provides real implementations for the stubs in `runtime.js`, making the instrumentation and the `Snapshot` polyfill actually do anything.  It also patches a large number of builtin APIs to do correct snapshotting/wrapping.  Note that the specifics of how patching is done are important to enable correct DCE - specifically, it must be done at the top level (rather than inside `$jscomp.polyfill`) to ensure that `RemoveUnusedCode` can actually see the calls at the right time.

The monkey-patches are guarded behind the `--instrument_async_context` compiler flag.  No behavior changes are expected unless that flag is enabled.

PiperOrigin-RevId: 643423125