Support tagged template literals as `@idgenerators`
This is superior as a syntax since it is syntactically enforced that the parameter must be a string literal.

When passed a template with parameters, the literal parts will be obfuscated but the substitutions will be preserved and passed to the function.

For this to work, idgenerator replacement needs to move to run before tranpilation so we can see the 'raw' ttl AST.

PiperOrigin-RevId: 669435000