Remove code when a variable is assigned to itself in an OR node.
For example, the following code would not be removed:

```
var a = a || {};
```

This is because `RemoveUnusedCode` would not recognize that the `a` in the LHS of the assignment was the **same variable** as the `a` in the LHS of the `OR` node.

This is because the `isEquivalentTo` method of Node does not take into account the context of the node. In this case, the `a` in the assignment is a variable declaration, while the `a` in the OR node is a reference to a variable.

Instead of using `isEquivalentTo` method, I have made `RemoveUnusedCode` do a simple string comparision to make sure the following two `NAME` nodes have the same name:
1) the node to which a value is being assigned
2) the value being assigned and the

PiperOrigin-RevId: 673473251