Refactor JSCompiler's `AsyncContext` polyfill and transpilation to use one runtime function instead of three.
The new version simplifies communication between the transpilation and the runtime library, and also produces smaller code.

Instead of separate enter/exit/reenter functions that create and operate on a token object, we now only have a `$jscomp.asyncContextEnter` function that returns a "contextSwap" closure to serve as both exit and reenter (depending on the presence of the second argument).  This function is also saved as `AsyncContext.Variable._JSC` for cross-binary interop.

PiperOrigin-RevId: 682437782