Fix generators to correctly exit the context on the initial invocation
The rough idea is that we add an extra optional parameter to the `$jscomp.asyncContextEnter` function to indicate that the outer context should start as undefined.  This is important so that the first reenter call within the actual generator body correctly captures the outer context rather than assuming it's a redundant reentry.

PiperOrigin-RevId: 688379938