Introduce "strictMissingRequire" diagnostics.
This check fills the holes in the current "missingRequire" check.  It reports when `goog.module` imports or references a legacy namespace indirectly through a namespace defined in separate file.

The three cases that were previously ignored in a `goog.module` were:

1.  When parent namespace is imported, but a namespace provided in another file is referenced through the parent object (a missing require).
   `const A = require('A');   ref(A.B);`

2.  When both namespace are imported but the namespace is referenced via the parent namespace object (wrong alias).
   `const A = require('A'); const B = require('B');   ref(A.B);`

3.  When the child namespace is referenced via a destructuring import of the parent namespace (bad import).
   ` const {B} = require('A');`

This introduces the check but it is disabled by default.

PiperOrigin-RevId: 691283195