Run 2 passes over the AST in RewriteCallerCodeLocation
The first pass `FindCallerLocationFunctions` will visit the param list of each function, find the params that are assigned to the default `goog.callerLocation()` function, and store these function names in a map.

The second pass `RewriteCallerLocationFunctionCalls` will visit each call node and check if it is a function call that requires a transformation. If it does, it will rewrite the call node to include the code location.

PiperOrigin-RevId: 692321563