Guard against `super.prop = x` during async method transpilation.
This previously caused a compiler crash.  It's still illegal, but now we report a clearer error message.

This is technically valid syntax, but it's rare.  The only known uses of super prop assignment are in overridden setters, which must not be async.  Rather than support an unused codepath, we instead just prohibit it, with a known workaround being to move the illegal syntax into an ordinary method that can then be called from the async method.

PiperOrigin-RevId: 698534353