Backoff SourceExcerptProvider from attempting to get source code of stub files when reporting errors.
When JSCompiler generates an error report, it tries to format the error message with the relevant source code snippet. To do that, it relies on a SourceExcerptProvider to give it the relevant snippet. The implementation for this SourceExcerptProvider interface is provided by the compiler.

This CL makes the SourceExcerptProvider (i.e. the compiler implementing it) backoff when the source file is a stub source file.

The sequence of calls is:
`ThreadSafeDelegatingErrorManager.generateReport()` --> `BasicErrorManager.generateReport()` -->  .. --> `LightweightMessageFormatter.formatError()` --> `SourceExcerptProvider$SourceExcerpt$1.get()` --> `compiler.getSourceLine()`

Also adds a unit test to check that compiler reports an error without crashing.

Fixes the crash reported in b/379868495 synced to CL 697886232 - http://sponge2/4ea3023b-17f3-45b4-adb4-de5ce0d1e864

PiperOrigin-RevId: 698904429