Report `declareIcuTemplate()` placeholders in non-canonical format
Placeholders are always formated in UPPER_SNAKE_CASE in XMB / XTB files
and in ICU template messages. Explicitly report an error if a
`declareIcuTemplate()` attempts to use a placeholder name in some other
format for its `example` or `original_code` maps.

NOTE: This change does not report errors for code that was accepted before.

It was already the case that you would get an error message if you tried
to refer to `lowerCamelCasePlaceholderName` in the `example` or
`original_code` part of a `declareIcuTemplate()` message, but the error
was confusing. It would say that the placeholder didn't exist,
even if you actually wrote "{lowerCamelCasePlaceholderName}" in the
template string.

This change also adds a `JsMessage` method for getting a list of the
placeholder names in canonical format.

PiperOrigin-RevId: 700030842