Add support for ICU templates with placeholders in JSCompiler.
This CL fixes a JSCompiler bug where `declareIcuTemplate` messages that have multiple parts are not translated to other languages.

ICU messages created using `declareIcuTemplate` can get stored into the XMB
file as multiple parts if necessary to record example or original code text,
but those parts are lost and only the original text template string is recorded
when we replace the original declarations with protected versions early in
compilation. This means we don't have the full information needed to
re-calculate the message ID when doing the final message replacement.

As a result we had a bug where the wrong ID was calculated and the translation
could not be found. This change solves this problem by explicitly recording the
correctly calculated message ID in the protected message.

PiperOrigin-RevId: 700089624