Move back maven_artifacts.bzl and THIRD_PARTY_NOTICES to root directory.
It is not a proper dependency direction (root to a test) for info that's required by actual builds and the release of the compiler.

Note that tests in license_check is currently broken - the files cannot be found due to the way it is wired in and pretty sure never worked (refers to "third_party" which doesn't exist, python has dependencies that are not wired in, etc). It is also not running as part of CIs nor referred by release process. So I couldn't completely validate what it was trying to do. However I ensured that it says in a way that dependencies are still wired and the targets are still buildable.

I may look into fixing the test later on but wanted to get things in a more meaningful shape first.

PiperOrigin-RevId: 703583197