Fix JSType equality bug leading to bad type errors with native module checks.
This is a partial fix for b/140763807. The underlying issue is that while the typed universe is still being created, before a type is ultimately marked "resolved", nominal types are compared solely based on name. The Closure type system allows unique types in different scopes to share a name, though.

This fixes that bug for the specific case of one or both names being in a different goog.module. It's not a general fix, but should be enough to prevent this bug from blocking native module typechecking for goog.module.

PiperOrigin-RevId: 705118303