Add a missing type argument to `IDBOpenDBRequest`
`IDBRequest` is parameterized, but the sub-interface IDBOpenDBRequest does not specify any type arguments for it.

From my read of the spec, you get a `IDBOpenRequest` from the result of [`IDBFactory#open`](https://developer.mozilla.org/en-US/docs/Web/API/IDBFactory/open) or [`IDBFactory#deleteDatabase`](https://developer.mozilla.org/en-US/docs/Web/API/IDBFactory/deleteDatabase). For the former the `result` (which is what the parameterized type represents)  will be the newly created `IDBDatabase` and in the latter case the result will be `null`.

PiperOrigin-RevId: 714064473