Introduce the node `SWITCH_BODY` to encapsulate the cases of a switch statement.
The cases of a switch are currently children of the `SWITCH` node, along with the switch condition.  This leads to some difficulties correctly modeling the new scope introduced by a switch. It holds all cases, but not the switch condition itself.

Fixes a bug where the compiler incorrectly reported a variable in a switch condition was an "early reference".

Introducing SWITCH_BODY makes it easier to model scope, and also in my opinion makes it easier to write code iterating over all cases in a switch - it just needs to look at all children of the SWITCH_BODY, instead of having to special case skipping the switch condition.

Why not use a BLOCK? various places in the compiler assume that BLOCK implies "children of this node can be statements" - if we used a BLOCK here, we'd have to update those places to say "children can be statements unless in a SWITCH".

PiperOrigin-RevId: 715051442