Add templated TArrayBuffer to TypedArray types
This keeps the definitions in sync with TS typings.  In the d.ts files, the type parameter has a
default type. Closure JS does not allow for default values in type parameters, but it allows a type reference without a type argument, in which case the type is {null}. Thus, we don't use the template type is not for anything.

PiperOrigin-RevId: 715123561