Declare global define aliases that J2clUtilGetDefineRewriterPass can read from
This will allow goog.defines declared in goog.modules to be consumable by J2CL's
System.getProperty infrastructure. Since these reads are out-of-scope of the
declaration, we need to provide some hook to link the two. To do this we define
a global variable to serve as a well-known alias to the define.

For the J2CL pass we can assume what the global variable is based on the define
name, so we just simply replace the getProperty references with the appropriate
global, if it exists.

PiperOrigin-RevId: 719370016