Add NON_LEGACY_GOOG_MODULE_REFERENCE warning to CheckMissingRequires.
This warnings on direct references of a goog.module namespace without goog.module.declareLegacyNamespace() in code.

This pattern already does not work at runtime - without out goog.module.declareLegacyNamespace(), the name of a goog.module is just an opaque identifier in code. This warning should provide stronger guarantees than the existing type system "inexistent property" errors, which are not always reliable.

Referencing a goog.module namespace in a type annotation still is allowed in scripts.

PiperOrigin-RevId: 726613351