Add correctness checks to J2clUtilGetDefineRewriterPass
This will enforce constraints on how we expect these calls to appear in the code
in order to properly optimize them. The primary thing we'll be looking for is
a corresponding `addSystemPropertyFromGoogDefine` call which help ensure that
the user wired things up correctly and we'll be replacing the code with the
correct define value.

PiperOrigin-RevId: 739187942