Refactor the conformance requirements merging logic to more clearly validate and merge requirements
This CL improves readability, and it will also allows us to more clearly validate that the extending configs correctly overrides the behavior field (`library_level_non_allowlisted_conformance_violations`) in the base configs if it's allowed to do so (e.g. report an error if both base config and extending config set a different value of the behavior field).

PiperOrigin-RevId: 744323441