Automate the opensource maven release of JsCompiler.
This change introduces a script that:
  - build the different bundles with bazel
  - extract all the files in a way they can be consumed by maven
  - use maven to package, sign and/or deploy and release all artifacts of every projects on sonatype.

The release number can be pass to the script trough the `RELEASE_NUM` env variable as before.

Pass the `--sonatype-auto-release` flag to the script to automatically release the projects on maven central (no manual intervention)

To run the script maven need to be installed on the host.
Maven will sign the artifacts using gpg. You need to either set your passphrase to the environment variable "MAVEN_GPG_PASSPHRASE" or in your local maven setting (~/.m2/settings.xml)
```
<properties>
  <gpg.passphrase>your_passphrase</gpg.passphrase>
</properties>
```

To setup the nexus sonatype credentions in maven, add a server section to your maven local settings (~/.m2/settings.xml):
```
 <servers>
    <server>
      <id>ossrh</id>
      <username>username</username>
      <password>username_password</password>
    </server>
  </servers>
```

An example of maven settings file: https://paste.googleplex.com/5921810779209728

PiperOrigin-RevId: 748144309