Always run PeepholeFoldConstants in the early peephole optimizations
Do not collapse `valueOf()` when `PeepholeFoldConstants` is run in `EarlyPeepholeOptimizations`, because `valueOf()` is a special function that should not be collapsed before `markPureFunctions`. `MarkPureFunctions` sees `valueOf()` calls as being side-effect free.

This change makes JSCompiler treat `false && something()` as dead code and delete it, which fixes an Angular bug where code of this pattern would not be deleted.

PiperOrigin-RevId: 762061405