Store the complete requirement within the violation object
#### Why?

To know whether a violation is a Boq Web violation, we compare violation's requirement (stored within the violation itself) with the requirements stored within the `libraryLevelNonAllowlistedConformanceViolationsBehavior` map (which is populated from the conformance config files).

However, the JSCompiler currently only stores the type, java_class and value fields of the requirements. This leads to incomplete comparison, and over-matching of requirements.

This CL enables us to accurately compare all fields of the requirements such as `rule_id` (except allowlists, etc) to ensure that the two requirements are truly equal.

PiperOrigin-RevId: 785961601