Add more debugging information to JSC_UNUSED_ARTIFICIAL_PURE_ANNOTATION
Note: this will only actually collect more data when 1) a given compilation actually enables JSC_UNUSED_ARTIFICIAL_PURE_ANNOTATION as a warning or error and 2) the compiler finds a function definition(s) that will cause an error.

If there wouldn't be any error/warning, this is a no-op & we don't use any more memory.

PiperOrigin-RevId: 788996377