Fix the isBoqWebViolation method to correctly find the Boq Web violations stored in the compiler's map
This CL is doing the same work as cl/783008709 except that cl/783008709 fixes the comparison in the checker tool whereas this CL is fixing the comparison in JSCompiler.

#### Why?
The existing code tries to look up the `Requirement`s in the map using `map.containsKey()`, which does not return true even for matching requirements if the underlying lists (allowlist, whitelists, values, etc) in the Requirement message are merged / sorted / shuffled. This comparison was fragile, and causes some violations to not getting matched, and therefore not get included in the library-level reports.

To fix, I'm comparing the relevant fields that can sufficiently ensure that two Requirements match each other, and ensuring that the order of items inside the "values" list does not influence the matching.

PiperOrigin-RevId: 794174450