Fix incorrect code generation for try...finally with break in a generator.
When a `break` statement appeared inside a `finally` block, the transpiler would generate an extra `break` statement, which would cause the `leaveFinallyBlock` call to be unreachable.

This change fixes the issue by introducing an `inFinallyBlock` flag to the `TranspilationContext`. This flag is used in `replaceBreakContinueWithJump` to prevent the extra `break` from being generated when inside a `finally` block.

PiperOrigin-RevId: 795630876