Fix covariant subtyping of ReadonlyArray & Array types
JSCompiler currently doesn't really treat ReadonlyArray as a covariant type unless it's comparing it to other ReadonlyArrays.

This has historically been fine because the main subtype of ReadonlyArray is regular Array, and it's usually a valid structural subtype anyway.

However, adding Array.prototype.toSpliced breaks this, because of its parameter typing; so to unblock good typings for ReadonlyArray.prototype.toSpliced, we can fix ReadonlyArray.

PiperOrigin-RevId: 809149023