Modify `getTypedAstDeserializer` to handle the case of empty chunks.
When a chunk is empty, the initialization logic adds a "fill" file to normalize the AST, however nothing exists to provide the "TypedAST" for that file.  We know however that these files are always empty and it is trivial to handle this like other synthetic files (which are already special cased).

PiperOrigin-RevId: 822742528