Ensure release workflows use the newly created release tag.
Modify the release workflow to pass the newly detected release tag to the test and publish jobs.
When we merged Merge `tag_release_version_workflow.yaml` into `release.yaml` , we forgot to pass the new created tag as github ref to use to run the test and build the maven release so the master bracnh was used as default github ref.

PiperOrigin-RevId: 842379848