Add test case showing JSCompiler transpilation of class inheritance within a local scope
We could optionally simplify the output by detecting that C's constructor does not return a value, and producing:

let D = function() {
  C.apply(this, arguments);
}

Today this detection only works on globals/module exports, not function local classes.

This probably doesn't matter for normal code, but @closureUnaware code is always wrapped in a function.

PiperOrigin-RevId: 846298004